Replaced tabs with spaces. Fixed bug so that it works well when the searched number appears more than once in the process
